Rename and standardize 'googlemaps' to 'google' across modules
Unify naming conventions by renaming 'googlemaps' references to 'google' in services, value objects, and repositories. Updated variable names from 'coords' to 'coord' for clarity and consistency. Adjusted file paths and templates to reflect these changes.
