+++ gmarker/domain/service/google.py
@@ -1,7 +1,7 @@
 import requests
 
-from gmarker.domain.repository.googlemaps import PlaceRepository
-from gmarker.domain.valueobject.googlemaps import PlaceVO, ReviewVO, RequestBody
+from gmarker.domain.repository.google import PlaceRepository
+from gmarker.domain.valueobject.google import PlaceVO, ReviewVO, RequestBody
 from gmarker.models import Place
 from lib.geo.valueobject.coord import GoogleMapsCoord
 
@@ -137,9 +137,9 @@ class GoogleMapsService:
 
             # 新しいPlaceを登録
             latlng = place_data.get("location")
-            coords = None
+            coord = None
             if latlng:
-                coords = GoogleMapsCoord(
+                coord = GoogleMapsCoord(
                     latitude=latlng.get("latitude"),
                     longitude=latlng.get("longitude"),
                 )
@@ -147,7 +147,7 @@ class GoogleMapsService:
                 Place(
                     place_id=place_id,
                     name=place_data.get("displayName", {}).get("text"),
-                    location=coords.to_str() if coords else None,
+                    location=coord.to_str() if coord else None,
                     rating=place_data.get("rating"),
                 )
             )
