+++ gmarker/urls.py
@@ -5,6 +5,6 @@ from gmarker.views import IndexView, CoordinateRegisterView
 app_name = "mrk"
 urlpatterns = [
     path("", IndexView.as_view(), name="index"),
-    path("coords/create/", CoordinateRegisterView.as_view(), name="coordinate_create"),
+    path("coord/create/", CoordinateRegisterView.as_view(), name="coordinate_create"),
     path("search/<int:search_code>", IndexView.as_view(), name="nearby_search"),
 ]
